Add contributor-friendly tests and documentation for mail-to-ticket-converter
